+++ src/tsn_adapters/common/trufnetwork/tn.py
@@ -1,19 +1,26 @@
+import hashlib
+import json
+from math import ceil
 import os
+from typing import Any, Optional, TypedDict
+
 import pandas as pd
+from prefect import flow, task
+from prefect.context import TaskRunContext
 import trufnetwork_sdk_c_bindings.exports as truf_sdk
 import trufnetwork_sdk_py.client as tn_client
-
 from trufnetwork_sdk_py.utils import generate_stream_id
-from math import ceil
-from prefect import flow, task
-from dotenv import load_dotenv
-from typing import  Optional, TypedDict
-from tsn_adapters.blocks.tn_access import UNUSED_INFINITY_RETRIES, TNAccessBlock, tn_special_retry_condition, StreamAlreadyExistsError
+
+from tsn_adapters.blocks.tn_access import (
+    UNUSED_INFINITY_RETRIES,
+    StreamAlreadyExistsError,
+    TNAccessBlock,
+    tn_special_retry_condition,
+)
+from tsn_adapters.utils.logging import get_logger_safe
 from tsn_adapters.utils.time_utils import date_string_to_unix
 from tsn_adapters.utils.tn_record import create_record_batches
-from tsn_adapters.utils.logging import get_logger_safe
 
-load_dotenv()
 
 @task(
     retries=UNUSED_INFINITY_RETRIES,
@@ -31,11 +38,7 @@ def task_insert_tsn_records(
 
 
 def insert_tsn_records(
-    stream_id: str,
-    records: pd.DataFrame,
-    client: tn_client.TNClient,
-    wait: bool = True,
-    records_per_batch: int = 300
+    stream_id: str, records: pd.DataFrame, client: tn_client.TNClient, wait: bool = True, records_per_batch: int = 300
 ):
     if len(records) == 0:
         print(f"No records to insert for stream {stream_id}")
@@ -46,8 +49,9 @@ def insert_tsn_records(
     records_dict = records.to_dict(orient="records")
     num_batches = ceil(len(records_dict) / records_per_batch)
     batches = create_record_batches(stream_id, records_dict, num_batches, records_per_batch)
-        
+
     client.batch_insert_records(batches, wait)
+
 
 @task(
     retries=UNUSED_INFINITY_RETRIES,
@@ -62,11 +66,9 @@ def task_insert_multiple_tsn_records(
 ):
     return insert_multiple_tsn_records(data, client, wait)
 
+
 def insert_multiple_tsn_records(
-    data: dict[str, pd.DataFrame],
-    client: tn_client.TNClient,
-    wait: bool = True,
-    records_per_batch: int = 300
+    data: dict[str, pd.DataFrame], client: tn_client.TNClient, wait: bool = True, records_per_batch: int = 300
 ):
     for stream_id, records in data.items():
         print(f"Processing stream {stream_id}...")
@@ -82,15 +84,18 @@ def insert_multiple_tsn_records(
             start_idx = batch_num * records_per_batch
             end_idx = start_idx + records_per_batch
             batch_records = records_dict[start_idx:end_idx]
-            
-            print(f"Processing batch {batch_num + 1}/{num_batches} for stream {stream_id} "
-                  f"(records {start_idx + 1}-{min(end_idx, total_records)})")
-            
+
+            print(
+                f"Processing batch {batch_num + 1}/{num_batches} for stream {stream_id} "
+                f"(records {start_idx + 1}-{min(end_idx, total_records)})"
+            )
+
             # Create and insert the batch
             batches = create_record_batches(stream_id, batch_records, batch_num + 1, records_per_batch)
             client.batch_insert_records(batches, wait)
-        
+
         print(f"Finished processing all batches for stream {stream_id}\n")
+
 
 """
 This task fetches all the records from the TSN for a given stream_id and data_provider
@@ -99,6 +104,8 @@ This task fetches all the records from the TSN for a given stream_id and data_pr
 - data_provider: the data provider to fetch the records from. Optional, if not provided, will use the one from the client
 - tsn_provider: the TSN provider to fetch the records from
 """  # noqa: E501
+
+
 @task(tags=["tn", "tn-read"])
 def task_get_all_tsn_records(
     stream_id: str, client: tn_client.TNClient, data_provider: Optional[str] = None
@@ -110,7 +117,9 @@ def get_all_tsn_records(
     stream_id: str, client: tn_client.TNClient, data_provider: Optional[str] = None
 ) -> pd.DataFrame:
     try:
-        recs = client.get_records(stream_id=stream_id, data_provider=data_provider, date_from=date_string_to_unix("1000-01-01"))
+        recs = client.get_records(
+            stream_id=stream_id, data_provider=data_provider, date_from=date_string_to_unix("1000-01-01")
+        )
     except:
         recs = []
 
@@ -195,7 +204,7 @@ def task_batch_deploy_streams(
 
 @task(
     name="Batch Filter Streams by Existence",
-    retries=UNUSED_INFINITY_RETRIES, # Read operations might not need as aggressive retries, adjust if necessary
+    retries=UNUSED_INFINITY_RETRIES,  # Read operations might not need as aggressive retries, adjust if necessary
     retry_delay_seconds=5,
     retry_condition_fn=tn_special_retry_condition(3),
     tags=["tn", "tn-read"],
@@ -215,19 +224,13 @@ def task_batch_filter_streams_by_existence(
         A list of stream locators matching the filter criteria.
     """
     logger = get_logger_safe()
-    logger.info(
-        f"Task: Batch filtering {len(locators)} streams by existence (return_existing: {return_existing})."
-    )
+    logger.info(f"Task: Batch filtering {len(locators)} streams by existence (return_existing: {return_existing}).")
     try:
-        filtered_locators = block.batch_filter_streams_by_existence(
-            locators=locators, return_existing=return_existing
-        )
+        filtered_locators = block.batch_filter_streams_by_existence(locators=locators, return_existing=return_existing)
         logger.info(f"Task: Found {len(filtered_locators)} matching streams after filtering.")
         return filtered_locators
     except Exception as e:
-        logger.error(
-            f"Task: Error during batch filtering of {len(locators)} streams: {e!s}", exc_info=True
-        )
+        logger.error(f"Task: Error during batch filtering of {len(locators)} streams: {e!s}", exc_info=True)
         raise
 
 
@@ -239,15 +242,15 @@ class FilterAndDeployResult(TypedDict):
 
 @task(
     name="Filter and Deploy Streams",
-    retries=UNUSED_INFINITY_RETRIES, # Relies on sub-task retries primarily
+    retries=UNUSED_INFINITY_RETRIES,  # Relies on sub-task retries primarily
     retry_delay_seconds=5,
-    retry_condition_fn=tn_special_retry_condition(2), # Fewer retries for the orchestrator
+    retry_condition_fn=tn_special_retry_condition(2),  # Fewer retries for the orchestrator
     tags=["tn", "tn-write", "tn-read"],
 )
 def task_filter_and_deploy_streams(
     block: TNAccessBlock,
     potential_definitions: list[tn_client.StreamDefinitionInput],
-    wait_for_deployment_tx: bool = False, # If True, task_batch_deploy_streams waits
+    wait_for_deployment_tx: bool = False,  # If True, task_batch_deploy_streams waits
 ) -> FilterAndDeployResult:
     """
     Filters a list of potential stream definitions for existence on TN,
@@ -295,14 +298,16 @@ def task_filter_and_deploy_streams(
         existent_locators = task_batch_filter_streams_by_existence(
             block=block, locators=locators_for_check, return_existing=True
         )
-        
+
         existent_sids = {str(loc["stream_id"]) for loc in existent_locators}
         logger.debug(f"Task: Found {len(existent_sids)} streams already on TN.")
 
     except Exception as e:
-        logger.error(f"Task: Error during batch stream existence check: {e!s}. Cannot proceed with this batch.", exc_info=True)
+        logger.error(
+            f"Task: Error during batch stream existence check: {e!s}. Cannot proceed with this batch.", exc_info=True
+        )
         # This is critical; if we can't check existence, we shouldn't deploy.
-        raise # Re-raise to let Prefect handle retry for this orchestrating task
+        raise  # Re-raise to let Prefect handle retry for this orchestrating task
 
     # 3. Determine streams to deploy and skipped streams
     definitions_to_deploy: list[tn_client.StreamDefinitionInput] = []
@@ -311,13 +316,17 @@ def task_filter_and_deploy_streams(
             skipped_sids_exist.append(sid)
         else:
             definitions_to_deploy.append(defn)
-    
-    logger.info(f"Task: {len(definitions_to_deploy)} streams to deploy, {len(skipped_sids_exist)} streams already exist.")
+
+    logger.info(
+        f"Task: {len(definitions_to_deploy)} streams to deploy, {len(skipped_sids_exist)} streams already exist."
+    )
 
     # 4. Deploy non-existent streams if any
     if definitions_to_deploy:
         try:
-            logger.debug(f"Task: Deploying {len(definitions_to_deploy)} streams (wait_for_tx: {wait_for_deployment_tx}).")
+            logger.debug(
+                f"Task: Deploying {len(definitions_to_deploy)} streams (wait_for_tx: {wait_for_deployment_tx})."
+            )
             # Call task directly
             tx_hash = task_batch_deploy_streams(
                 block=block, definitions=definitions_to_deploy, wait=wait_for_deployment_tx
@@ -327,7 +336,7 @@ def task_filter_and_deploy_streams(
         except Exception as e:
             logger.error(f"Task: Error during batch stream deployment: {e!s}.", exc_info=True)
             # This is critical for the streams that were attempted.
-            raise # Re-raise
+            raise  # Re-raise
     else:
         logger.info("Task: No new streams to deploy.")
 
@@ -339,6 +348,7 @@ def task_filter_and_deploy_streams(
 
 
 if __name__ == "__main__":
+
     @flow(log_prints=True)
     def test_get_all_tsn_records():
         client = tn_client.TNClient(url=os.environ["TSN_PROVIDER"], token=os.environ["TSN_PRIVATE_KEY"])
@@ -354,22 +364,68 @@ if __name__ == "__main__":
         client.deploy_stream(stream_id)
 
         # insert records into test stream
-        insert_tsn_records(stream_id, pd.DataFrame({
-            "date": [date_string_to_unix("2023-01-01"), date_string_to_unix("2023-01-02")],
-            "value": [10.5, 20.3]
-        }), client)
-        insert_multiple_tsn_records({
-            stream_id: pd.DataFrame({
-                "date": [date_string_to_unix("2023-01-03")],
-                "value": [12]
-            })
-        }, client)
+        insert_tsn_records(
+            stream_id,
+            pd.DataFrame(
+                {"date": [date_string_to_unix("2023-01-01"), date_string_to_unix("2023-01-02")], "value": [10.5, 20.3]}
+            ),
+            client,
+        )
+        insert_multiple_tsn_records(
+            {stream_id: pd.DataFrame({"date": [date_string_to_unix("2023-01-03")], "value": [12]})}, client
+        )
 
         # get records
         recs = get_all_tsn_records(stream_id, client)
         print(recs)
-        
+
         # cleanup
         client.destroy_stream(stream_id)
 
     test_get_all_tsn_records()
+
+# --- Custom Cache Key Function ---
+
+
+def filter_existence_cache_key(
+    context: TaskRunContext,  # Prefect provides context
+    parameters: dict[str, Any],  # Prefect provides parameters
+) -> str:
+    """
+    Generates a cache key for task_batch_filter_streams_by_existence
+    based on a stable representation of locators and the return_existing flag.
+    Raises KeyError if required parameters are missing.
+    """
+    try:
+        locators = parameters["locators"]  # Direct access, will raise KeyError if missing
+        return_existing = parameters["return_existing"]  # Direct access
+    except KeyError as e:
+        # Fail fast with an informative error
+        raise KeyError(
+            f"Cache key generation failed: Missing expected parameter '{e.args[0]}' for task referenced by "
+            f"TaskRunContext '{context.task_run.name if context and context.task_run else 'Unknown'}'"
+        ) from e
+
+    # Proceed with sorting and hashing only if parameters are present
+    try:
+        # Ensure keys exist and handle potential None values robustly
+        stable_locators = sorted(
+            [(str(loc.get("stream_id", "")), str(loc.get("data_provider", ""))) for loc in locators]
+        )
+    except Exception as e:
+        # Log error if locators aren't as expected, fall back to less specific key
+        # get_run_logger() might not be available outside task/flow context, use basic print/log
+        print(f"WARNING: Could not generate stable locator list for caching: {e}. Using basic key.")
+        stable_locators = []  # Avoid hashing potentially unstable input on error
+
+    # Combine with return_existing flag
+    key_data = {
+        "locators_repr": stable_locators,  # Use a different key name from param
+        "return_existing": return_existing,
+    }
+
+    # Use json dumps with sort_keys for a stable string representation, then hash
+    # Using sha256 for robustness against potential hash collisions with large inputs
+    hasher = hashlib.sha256()
+    hasher.update(json.dumps(key_data, sort_keys=True).encode("utf-8"))
+    return hasher.hexdigest()
