+++ tests/fixtures/test_trufnetwork.py
@@ -1,5 +1,6 @@
 from collections.abc import Generator
 from dataclasses import dataclass, field
+from functools import wraps
 import json
 import logging
 import os
@@ -7,10 +8,10 @@ import shutil
 import signal
 import subprocess
 import time
-from typing import Any, Optional
-from unittest.mock import Mock
+from typing import Any, Callable, Optional
+from unittest.mock import Mock, patch
 
-from prefect import Task
+from prefect import Task, task as original_prefect_task
 from prefect.client.orchestration import get_client
 from prefect.client.schemas.actions import GlobalConcurrencyLimitCreate
 from prefect.exceptions import ObjectNotFound
@@ -400,100 +401,98 @@ def term_handler():
 
 @pytest.fixture(scope="session", autouse=True)
 def disable_prefect_retries():
-    from importlib import import_module
-    from typing import Any
-    from unittest.mock import patch
+    """
+    Patches prefect.task globally for the test session to ensure all tasks
+    are created with retries=0 and no caching, and that subsequent
+    .with_options() calls also enforce this.
+    """
 
-    from prefect import task as original_task
+    # This function takes a Task instance and modifies it
+    def apply_test_options_to_task(task_instance: Task[Any, Any]) -> Task[Any, Any]:
+        if not isinstance(task_instance, Task): # type: ignore[unreachable]
+            # Should not happen if original_prefect_task behaves as expected
+            return task_instance
 
-    # Patch task retries by modifying the task options directly
-    def patch_task_options(task_fn: Task[Any, Any]) -> Task[Any, Any]:
         # Always override retries, cache_key_fn, and cache_expiration
         # regardless of how the task was created
-        if hasattr(task_fn, "with_options"):
-            patched_task = task_fn.with_options(
-                retries=0, cache_key_fn=None, cache_expiration=None, retry_condition_fn=None
-            )
+        # print(f"Patching options for task: {getattr(task_instance, 'name', 'Unnamed Task')}") # For debugging
+        patched_task = task_instance.with_options(
+            retries=0,
+            cache_key_fn=None,
+            cache_expiration=None,
+            retry_condition_fn=None,
+            # You might want to set a specific task_run_name for easier test debugging
+            # task_run_name=f"{task_instance.name}-test-run" if task_instance.name else "unnamed-task-test-run"
+        )
 
-            # Also patch the with_options method of the returned task to ensure
-            # any subsequent calls also have these options overridden
-            original_with_options = patched_task.with_options
+        # Also patch the with_options method of the returned task to ensure
+        # any subsequent calls also have these options overridden
+        original_with_options = patched_task.with_options
 
-            # Use a simple function that ignores type checking
-            def ensure_no_retries_or_cache(**kwargs: Any) -> Any:
-                # Force these options regardless of what was passed
-                kwargs["retries"] = 0
-                kwargs["cache_key_fn"] = None
-                kwargs["cache_expiration"] = None
-                kwargs["retry_condition_fn"] = None
-                return original_with_options(**kwargs)  # type: ignore
+        @wraps(original_with_options)
+        def ensure_no_retries_or_cache_on_with_options(*args: Any, **kwargs: Any) -> Task[Any, Any]:
+            # Force these options regardless of what was passed
+            # The first arg to with_options can be a function for state handlers,
+            # so we need to be careful if we're modifying args.
+            # It's safer to just update kwargs.
+            kwargs["retries"] = 0
+            kwargs["cache_key_fn"] = None
+            kwargs["cache_expiration"] = None
+            kwargs["retry_condition_fn"] = None
+            return original_with_options(*args, **kwargs)
 
-            # Use setattr to avoid type checking issues
-            setattr(patched_task, "with_options", ensure_no_retries_or_cache)
+        # Use setattr to avoid type checking issues if any, or direct assignment
+        # setattr(patched_task, "with_options", ensure_no_retries_or_cache_on_with_options)
+        patched_task.with_options = ensure_no_retries_or_cache_on_with_options # type: ignore[assignment]
 
-            return patched_task
-        return task_fn
+        return patched_task
 
-    # All tasks with retries and their import paths
-    tasks_to_patch = [
-        # FMP Historical Flow
-        "tsn_adapters.flows.fmp.historical_flow.fetch_historical_data",
-        "tsn_adapters.flows.fmp.historical_flow.get_earliest_data_date",
-        # Stream Deploy Flow
-        "tsn_adapters.flows.stream_deploy_flow.check_deploy_stream",
-        # Primitive Source Descriptor
-        "tsn_adapters.blocks.primitive_source_descriptor.get_descriptor_from_url",
-        "tsn_adapters.blocks.primitive_source_descriptor.get_descriptor_from_github",
-        # FMP Real Time Flow
-        "tsn_adapters.flows.fmp.real_time_flow.fetch_quotes_for_batch",
-        # Argentina Task Wrappers
-        "tsn_adapters.tasks.argentina.task_wrappers.task_create_reconciliation_strategy",
-        "tsn_adapters.tasks.argentina.task_wrappers.task_create_sepa_provider",
-        "tsn_adapters.tasks.argentina.task_wrappers.task_create_stream_fetcher",
-        "tsn_adapters.tasks.argentina.task_wrappers.task_create_transformer",
-        "tsn_adapters.tasks.argentina.task_wrappers.task_dates_already_processed",
-        "tsn_adapters.tasks.argentina.task_wrappers.task_determine_needed_keys",
-        "tsn_adapters.tasks.argentina.task_wrappers.task_get_and_transform_data",
-        "tsn_adapters.tasks.argentina.task_wrappers.task_get_data_for_date",
-        "tsn_adapters.tasks.argentina.task_wrappers.task_get_latest_records",
-        "tsn_adapters.tasks.argentina.task_wrappers.task_get_now_date",
-        "tsn_adapters.tasks.argentina.task_wrappers.task_get_streams",
-        "tsn_adapters.tasks.argentina.task_wrappers.task_insert_data",
-        "tsn_adapters.tasks.argentina.task_wrappers.task_load_category_map",
-        "tsn_adapters.tasks.argentina.task_wrappers.task_transform_data",
-        # Argentina Preprocess Flow
-        "tsn_adapters.tasks.argentina.flows.preprocess_flow.process_raw_data",
-        "tsn_adapters.tasks.argentina.task_wrappers.task_get_and_transform_data",
-        "tsn_adapters.tasks.argentina.task_wrappers.task_get_now_date",
-        "tsn_adapters.tasks.argentina.task_wrappers.task_dates_already_processed",
-        # TN Access
-        "tsn_adapters.blocks.tn_access.task_wait_for_tx",
-        "tsn_adapters.blocks.tn_access.task_insert_and_wait_for_tx",
-        "tsn_adapters.blocks.tn_access._task_only_batch_insert_records",
-        "tsn_adapters.blocks.tn_access.task_split_and_insert_records",
-        # TN Common
-        "tsn_adapters.common.trufnetwork.tn.task_insert_tsn_records",
-        "tsn_adapters.common.trufnetwork.tn.task_deploy_primitive",
-        "tsn_adapters.common.trufnetwork.tn.task_get_all_tsn_records",
-        # GSheet Tasks
-        "tsn_adapters.tasks.gsheet.task_read_gsheet",
-        # Argentina Preprocess Flow
-        "tsn_adapters.tasks.argentina.flows.preprocess_flow.task_list_available_dates",
-    ]
+    # This is the replacement for the prefect.task decorator
+    def patched_task_decorator_factory(*decorator_args: Any, **decorator_kwargs: Any) -> Any:
+        """
+        This function replaces `prefect.task`.
+        It handles both `@task` and `@task(...)` usage.
+        """
+        # Case 1: @task (used without arguments, decorating a function directly)
+        # e.g. @task
+        #      def my_fn(): ...
+        if len(decorator_args) == 1 and callable(decorator_args[0]) and not decorator_kwargs:
+            original_func = decorator_args[0]
+            # Create the task instance using the original decorator
+            task_instance = original_prefect_task(original_func)
+            # Apply our modifications
+            return apply_test_options_to_task(task_instance)
 
-    # Patch the task decorator to apply our options
-    with patch("prefect.task", side_effect=original_task) as mock_task:
-        for import_path in tasks_to_patch:
-            # Split the import path into module path and attribute name
-            module_path, attr_name = import_path.rsplit(".", 1)
-            # Import the module and get the task function
-            module = import_module(module_path)
-            task_fn = getattr(module, attr_name)
-            # Patch the task
-            mock_task.return_value = patch_task_options(task_fn)
-            patch(import_path, new=patch_task_options(task_fn)).start()
+        # Case 2: @task(...) (used with arguments, returns a decorator)
+        # e.g. @task(name="My Task", retries=3)
+        #      def my_fn(): ...
+        # Or direct call: my_task = task(my_fn, name="My Task")
+        else:
+            # It's either @task(options) or task(fn, options)
+            # Call original_prefect_task to get either the task or the inner decorator
+            task_or_inner_decorator = original_prefect_task(*decorator_args, **decorator_kwargs)
 
-    yield
+            if isinstance(task_or_inner_decorator, Task):
+                # This happens if called like: task(fn, name="foo")
+                return apply_test_options_to_task(task_or_inner_decorator)
+            elif callable(task_or_inner_decorator):
+                # This happens if called like: @task(name="foo")
+                # task_or_inner_decorator is now the thing that will take the function
+                @wraps(task_or_inner_decorator) # Preserves signature of the inner decorator
+                def actual_decorator_to_apply_options(fn: Callable[..., Any]) -> Task[Any, Any]:
+                    task_instance = task_or_inner_decorator(fn)
+                    return apply_test_options_to_task(task_instance)
+                return actual_decorator_to_apply_options
+            else:
+                # Should not be reached with standard prefect.task usage
+                return task_or_inner_decorator
+
+
+    # Patch prefect.task globally. All imports of prefect.task thereafter will get our version.
+    # This needs to happen before any of your tasks are imported and defined.
+    # autouse=True, scope="session" for the fixture helps ensure this.
+    with patch("prefect.task", new=patched_task_decorator_factory):
+        yield
 
 
 @pytest.fixture(scope="function")
